+++ chat-widget/src/components/chatbuttonstateful/ChatButtonStateful.tsx
@@ -32,7 +32,7 @@ export const ChatButtonStateful = (props: IChatButtonStatefulParams) => {
         titleText: "Let's Chat!",
         subtitleText: "We're online.",
         hideNotificationBubble: buttonProps?.controlProps?.hideNotificationBubble === true || state.appStates.isMinimized === false,
-        unreadMessageCount: state.appStates.unreadMessageCount ? (state.appStates.unreadMessageCount > Constants.maximumUnreadMessageCount ? Constants.maximumUnreadMessageCount.toString() + "+" : state.appStates.unreadMessageCount.toString()) : "0",
+        unreadMessageCount: state.appStates.unreadMessageCount ? (state.appStates.unreadMessageCount > Constants.maximumUnreadMessageCount ? props.buttonProps?.controlProps?.largeUnreadMessageString : state.appStates.unreadMessageCount.toString()) : "0",
         onClick: async () => {
             TelemetryHelper.logActionEvent(LogLevel.INFO, {
                 Event: TelemetryEvent.LCWChatButtonClicked
@@ -48,6 +48,7 @@ export const ChatButtonStateful = (props: IChatButtonStatefulParams) => {
                 await startChat();
             }
         },
+        unreadMessageString: props.buttonProps?.controlProps?.unreadMessageString,
         ...buttonProps?.controlProps,
     };
 
@@ -63,6 +64,7 @@ export const ChatButtonStateful = (props: IChatButtonStatefulParams) => {
                 dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.OutOfOffice });
             }
         },
+        unreadMessageString: props.buttonProps?.controlProps?.unreadMessageString,
         ...outOfOfficeButtonProps?.controlProps
     };